// frontend/src/api/generated/models/articleDetailResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * My API
 * OpenAPI spec version: 1.0.0
 */

export interface ArticleDetailResponse {
  /** 文章内容 */
  content: string;
  /** 创建时间 */
  createdAt: string;
  /** 简介描述 */
  description: string;
  /** ID */
  id: number;
  /** 标题 */
  title: string;
  /** 文章类型ID */
  typeId: number;
  /** 文章类型名称 */
  typeName: string;
}